// frontend/src/api/model/updateHostelDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Student Welfare Platform API
 * The API documentation for the Student Welfare Platform
 * OpenAPI spec version: 1.0
 */
import type { CreateHostelResourceDto } from './createHostelResourceDto';

export interface UpdateHostelDto {
  name?: string;
  location?: string;
  description?: string;
  resources?: CreateHostelResourceDto[];
}